Moved interfaces to another file and added barrel export.
