+++ src/aurelia-store.ts
@@ -31,3 +31,4 @@ export * from "./history";
 export * from "./middleware";
 export * from "./logging";
 export * from "./decorator";
+export * from "./devtools";
